Rename directory from VideoPlayer to Video
